src/app/requirements.txt
# 📁 requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
chromadb
pydantic
typing-extensions
requests
aiofiles
python-jose
passlib
Pillow
pandas
numpy
scipy